Successfully implemented the backend.
Getting hotels based on range is now functional

+++ hotelmanager/src/main/java/fts/intern/hotelmanager/dto/FeedbackDto.java
@@ -1,17 +1,13 @@
 package fts.intern.hotelmanager.dto;
 
-
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-
 public class FeedbackDto {
-    private Long id;
-    private Long hotel_id;
+    private int id;
+    private int hotel_id;
     private Double rating;
     private String comment;
-
 }
-
